backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
python-multipart==0.0.20
python-dotenv==1.0.1
sqlalchemy==2.0.36
greenlet==3.1.1
pillow==11.0.0
